Add a configuration menu to replace call to LADCONF.COM
No attempt has been made to mimic the appearance or behavior of
LADCONF.COM, only to generate a LADDER.DAT containing the required
fields.

The old code that executed LADCONF.COM could not be used, as not only
is it a CP/M executable, but it was being run using a method not
supported by FPC.  That code could have potentially been updated to
invoke it in a way supported by FPC and via a CP/M emulator, but this
is simpler.